python-fastapi-chatbot/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
openai==1.61.1
httpx==0.28.1
python-dotenv==1.0.1
jinja2==3.1.5
